qwen-topstepx-bundle/ml/requirements.txt
lightgbm
pandas
numpy
scikit-learn
joblib
pyarrow
pydantic
tqdm
torch>=2.0.0
torchvision>=0.15.0
alpaca-trade-api>=3.0.0
python-dotenv>=1.0.0
chronos-forecasting>=1.4.0
transformers>=4.40.0
